requirements.txt
torch
pyro-ppl
matplotlib~=3.4.2
seaborn~=0.11.1
numpy~=1.20.3
pytest~=6.2.4
Sphinx
torchvision~=0.8.2
scikit-learn~=0.24.2
tqdm~=4.61.1
setuptools~=49.6.0